c8d/list: Extract fake service helpers
Extract to a separate file as they're also used by other files in this
package.
